requirements.txt
python-telegram-bot==21.5
fpdf2==2.7.9
groq==0.9.0
